pyproject.toml
[project]
name = "pymander"
version = "0.1.0"
description = "16s Narrative Intelligence Platform"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115,<1",
    "uvicorn[standard]>=0.34,<1",
    "pydantic>=2.10,<3",
    "pydantic-settings>=2.7,<3",
    "sqlalchemy[asyncio]>=2.0,<3",
    "asyncpg>=0.30,<1",
    "alembic>=1.14,<2",
    "aiokafka>=0.12,<1",
    "redis[hiredis]>=5.2,<6",
    "structlog>=24.4,<26",
    "httpx>=0.28,<1",
    "pgvector>=0.3,<1",
    # Phase 2: Ingestion
    "asyncpraw>=7.8,<8",
    "feedparser>=6.0,<7",
    "trafilatura>=2.0,<3",
    "aiohttp>=3.11,<4",
    # Phase 2: Intelligence pipeline
    "sentence-transformers>=3.3,<4",
    "hdbscan>=0.8,<1",
    "qdrant-client>=1.12,<2",
    "numpy>=1.26,<3",
    # Phase 2: Image hashing
    "imagehash>=4.3,<5",
    "Pillow>=11,<12",
    # Phase 2: Narrative validation
    "anthropic>=0.42,<1",
    # Phase 2: Dashboard
    "websockets>=14,<15",
    # Phase 3: Network graph
    "neo4j>=5.24,<6",
    "networkx>=3.2,<4",
    # Phase 3: Identity resolution
    "python-Levenshtein>=0.26,<1",
    # Phase 3: Additional adapters
    "tweepy>=4.14,<5",
    "telethon>=1.37,<2",
    "google-api-python-client>=2.150,<3",
    "youtube-transcript-api>=0.6,<1",
    "pytrends>=4.9,<5",
    # Phase 3: Behavioral signals
    "scipy>=1.11,<2",
    "dtw-python>=1.5,<2",
    # Phase 4: API auth
    "pyjwt>=2.9,<3",
    "passlib[bcrypt]>=1.7,<2",
    # Phase 4: Briefings
    "jinja2>=3.1,<4",
    # Phase 4: SSE
    "sse-starlette>=2.0,<3",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3,<9",
    "pytest-asyncio>=0.25,<1",
    "pytest-cov>=6,<7",
    "ruff>=0.9,<1",
    "faker>=33,<40",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/pymander"]

[tool.ruff]
target-version = "py311"
src = ["src"]
line-length = 99

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "SIM"]
ignore = ["B008"]  # FastAPI Depends() in defaults is idiomatic
per-file-ignores = {"src/pymander/narrative/validator.py" = ["E501"], "src/pymander/intelligence/briefings.py" = ["E501"], "src/pymander/intelligence/coordination.py" = ["E501"], "src/pymander/network/analytics.py" = ["E501"]}

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff.lint.isort]
known-first-party = ["pymander"]
